Package rename, simple web controller test added
