+++ pom.xml
@@ -51,6 +51,11 @@
             <artifactId>spring-boot-starter-security</artifactId>
             <version>${spring.boot.version}</version>
         </dependency-->
+        <dependency>
+            <groupId>org.springframework.boot</groupId>
+            <artifactId>spring-boot-starter-test</artifactId>
+            <version>${spring.boot.version}</version>
+        </dependency>
 
         <dependency>
             <groupId>org.springdoc</groupId>
@@ -70,7 +75,6 @@
             <version>2.1.214</version>
             <scope>runtime</scope>
         </dependency>
-
 
     </dependencies>
