+++ src/main/java/edu/spring_practice/Main.java
@@ -1,4 +1,4 @@
-package edu.spring_test;
+package edu.spring_practice;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
