+++ src/main/java/edu/spring_practice/jpa/entities/Student.java
@@ -1,10 +1,14 @@
-package edu.spring_test.jpa.entities;
+package edu.spring_practice.jpa.entities;
 
 import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
 import lombok.Data;
+import lombok.NoArgsConstructor;
 
 @Data
 @Entity
+@AllArgsConstructor
+@NoArgsConstructor
 public class Student {
 
     @Id
