+++ src/main/java/edu/spring_practice/jpa/repositories/CourseRepository.java
@@ -1,6 +1,6 @@
-package edu.spring_test.jpa.repositories;
+package edu.spring_practice.jpa.repositories;
 
-import edu.spring_test.jpa.entities.Course;
+import edu.spring_practice.jpa.entities.Course;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
