+++ src/main/java/edu/spring_practice/jpa/repositories/StudentRepository.java
@@ -1,6 +1,6 @@
-package edu.spring_test.jpa.repositories;
+package edu.spring_practice.jpa.repositories;
 
-import edu.spring_test.jpa.entities.Student;
+import edu.spring_practice.jpa.entities.Student;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
