+++ src/test/java/edu/spring_practice/controllers/SimpleControllerTest.java
@@ -1,0 +1,28 @@
+package edu.spring_practice.controllers;
+
+import lombok.SneakyThrows;
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.http.MediaType;
+import org.springframework.test.web.servlet.MockMvc;
+import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
+import org.springframework.test.web.servlet.result.MockMvcResultMatchers;
+
+@SpringBootTest
+@AutoConfigureMockMvc
+public class SimpleControllerTest {
+
+    @Autowired
+    private MockMvc mock_mvc;
+
+    @Test
+    @SneakyThrows
+    public void testSum() {
+        mock_mvc.perform(MockMvcRequestBuilders.get("/sum").param("a", "3").param("b", "2").
+                contentType(MediaType.APPLICATION_JSON)).andExpect(MockMvcResultMatchers.status().isOk()).
+                andExpect(MockMvcResultMatchers.content().string("5"));
+    }
+
+}
